backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
pymysql==1.1.0
psycopg2-binary==2.9.9
pymongo==4.6.1
CouchDB==1.2
cryptography==42.0.5
pandas==2.2.0
tqdm==4.66.2
openai==1.12.0
pydantic==2.6.1
httpx==0.27.0
